+++ test/lib/functions/vUrlAudio.test.ts
@@ -1,21 +1,21 @@
 'use strict';
 
-import {describe, it}   from 'node:test';
-import * as assert      from 'node:assert/strict';
-import guid             from '@valkyriestudios/utils/hash/guid';
-import CONSTANTS        from '../../constants';
-import Validator        from '../../../lib';
+import {describe, it}       from 'node:test';
+import * as assert          from 'node:assert/strict';
+import guid                 from '@valkyriestudios/utils/hash/guid';
+import CONSTANTS            from '../../constants';
+import Validator            from '../../../lib';
 import {
     AUDIO_EXTENSIONS,
     VIDEO_EXTENSIONS,
     IMAGE_EXTENSIONS,
 } from '../../../lib/functions/vUrlExtensions';
 
-describe('vUrlImage', () => {
-    it('Should validate a url string that is a valid url and an image url as valid', () => {
-        const v = new Validator({a: 'url_img'});
+describe('vUrlAudio', () => {
+    it('Should validate a url string that is a valid url and a audio url as valid', () => {
+        const v = new Validator({a: 'url_aud'});
 
-        IMAGE_EXTENSIONS.forEach(ext => {
+        AUDIO_EXTENSIONS.forEach(ext => {
             for (const el of [
                 `http://foo.com/blah_blah/${guid()}.${ext}`,
                 `http://foo.com/blah_blah_(wikipedia).${ext}`,
@@ -55,10 +55,10 @@ describe('vUrlImage', () => {
         });
     });
 
-    it('Should validate a url string that is a valid media url but not an image url as invalid', () => {
-        const v = new Validator({a: 'url_img'});
+    it('Should validate a url string that is a valid media url but not a audio url as invalid', () => {
+        const v = new Validator({a: 'url_aud'});
 
-        for (const ext of [...VIDEO_EXTENSIONS, ...AUDIO_EXTENSIONS]) {
+        for (const ext of [...VIDEO_EXTENSIONS, ...IMAGE_EXTENSIONS].filter(el => el !== 'ogg')) {
             for (const el of [
                 `http://foo.com/blah_blah/${guid()}.${ext}`,
                 `http://foo.com/blah_blah_(wikipedia).${ext}`,
@@ -97,15 +97,15 @@ describe('vUrlImage', () => {
                     is_valid: false,
                     count: 1,
                     errors: {
-                        a: [{msg: 'url_img', params: []}],
+                        a: [{msg: 'url_aud', params: []}],
                     },
                 });
             }
         }
     });
 
-    it('Should not validate a url string that is a valid url but not an image url as valid', () => {
-        const v = new Validator({a: 'url_img'});
+    it('Should not validate a url string that is a valid url but not a audio url as valid', () => {
+        const v = new Validator({a: 'url_aud'});
 
         for (const el of [
             'http://foo.com/blah_blah',
@@ -160,21 +160,21 @@ describe('vUrlImage', () => {
             'https://valkyriestudios.com/?a=1&b=2#eweqweqwewqeqweweqweqweqweqwewopqieqwoeqweipqwoeiqweiqwpoeiqwieqwopeiqwpooeqwieiqwoeiqwoeiqweiqwopeiqwpeiqweiqweoiwqopepwoqiepqwoieoiwqeipqwoeiowqieoipoeiwqoiepowiqoieqw&ewquie', // eslint-disable-line max-len
             'https://valkyriestudios.com:3000/?a=1&b=2#eweqweqwewqeqweweqweqweqweqwewopqieqwoeqweipqwoeiqweiqwpoeiqwieqwopeiqwpooeqwieiqwoeiqwoeiqweiqwopeiqwpeiqweiqweoiwqopepwoqiepqwoieoiwqeipqwoeiowqieoipoeiwqoiepowiqoieqw&ewquie', // eslint-disable-line max-len
             'http://x.comddfsdfsdf.', // Trailing dots in tlds are valid
-            'http://bla/.jpg',
+            'https://bla/.mp3',
         ]) {
             const evaluation = v.validate({a: el});
             assert.deepEqual(evaluation, {
                 is_valid: false,
                 count: 1,
                 errors: {
-                    a: [{msg: 'url_img', params: []}],
+                    a: [{msg: 'url_aud', params: []}],
                 },
             });
         }
     });
 
     it('Should not validate other types as valid urls', () => {
-        const v = new Validator({a: 'url_img'});
+        const v = new Validator({a: 'url_aud'});
 
         for (const el of CONSTANTS.NOT_STRING_WITH_EMPTY) {
             const evaluation = v.validate({a: el});
@@ -182,14 +182,14 @@ describe('vUrlImage', () => {
                 is_valid: false,
                 count: 1,
                 errors: {
-                    a: [{msg: el === undefined ? 'not_found' : 'url_img', params: []}],
+                    a: [{msg: el === undefined ? 'not_found' : 'url_aud', params: []}],
                 },
             });
         }
     });
 
-    it('Should not validate strings that are image urls after trimming as valid', () => {
-        const v = new Validator({a: 'url_img'});
+    it('Should not validate strings that are audio urls after trimming as valid', () => {
+        const v = new Validator({a: 'url_aud'});
 
         for (const el of [
             'https://www.myfancylink.com/123.jpg   ',
@@ -201,14 +201,14 @@ describe('vUrlImage', () => {
                 is_valid: false,
                 count: 1,
                 errors: {
-                    a: [{msg: 'url_img', params: []}],
+                    a: [{msg: 'url_aud', params: []}],
                 },
             });
         }
     });
 
     it('Should not validate improper strings as valid urls', () => {
-        const v = new Validator({a: 'url_img'});
+        const v = new Validator({a: 'url_aud'});
 
         for (const el of [
             'http://',
@@ -254,7 +254,7 @@ describe('vUrlImage', () => {
                 is_valid: false,
                 count: 1,
                 errors: {
-                    a: [{msg: 'url_img', params: []}],
+                    a: [{msg: 'url_aud', params: []}],
                 },
             });
         }